backend/requirements.txt
fastapi==0.115.14
uvicorn==0.35.0
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.72
langchain-openai==0.3.27
langchain_mcp_adapters==0.1.9
mcp_use==1.3.6
pymongo==4.13.2
python-dotenv==1.1.1
httpx==0.28.1
langgraph==0.5.2
pydantic==2.10.6
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
python-multipart==0.0.20